handle logic buttons onclick in Non Fragment
